rcmnd-service/requirements.txt
Authlib==1.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
idna==3.4
flask
pycparser==2.21
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.4
validators==0.21.0
weaviate-client==3.23.2
